fix url

+++ routes/admin/hoi-dong.php
@@ -4,11 +4,11 @@ use App\Http\Controllers\Admin\HoiDongController;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware(['auth', 'role:admin'])->group(function () {
-    Route::get('/hoi-dong', [HoiDongController::class, 'index'])->name('admin.hoi-dong.index');
-    Route::get('/hoi-dong/create', [HoiDongController::class, 'create'])->name('admin.hoi-dong.create');
-    Route::post('/hoi-dong', [HoiDongController::class, 'store'])->name('admin.hoi-dong.store');
-    Route::get('/hoi-dong/{hoiDong}', [HoiDongController::class, 'show'])->name('admin.hoi-dong.show');
-    Route::get('/hoi-dong/{hoiDong}/edit', [HoiDongController::class, 'edit'])->name('admin.hoi-dong.edit');
-    Route::put('/hoi-dong/{hoiDong}', [HoiDongController::class, 'update'])->name('admin.hoi-dong.update');
-    Route::delete('/hoi-dong/{hoiDong}', [HoiDongController::class, 'destroy'])->name('admin.hoi-dong.destroy');
+    Route::get('/admin/hoi-dong', [HoiDongController::class, 'index'])->name('admin.hoi-dong.index');
+    Route::get('/admin/hoi-dong/create', [HoiDongController::class, 'create'])->name('admin.hoi-dong.create');
+    Route::post('/admin/hoi-dong', [HoiDongController::class, 'store'])->name('admin.hoi-dong.store');
+    Route::get('/admin/hoi-dong/{hoiDong}', [HoiDongController::class, 'show'])->name('admin.hoi-dong.show');
+    Route::get('/admin/hoi-dong/{hoiDong}/edit', [HoiDongController::class, 'edit'])->name('admin.hoi-dong.edit');
+    Route::put('/admin/hoi-dong/{hoiDong}', [HoiDongController::class, 'update'])->name('admin.hoi-dong.update');
+    Route::delete('/admin/hoi-dong/{hoiDong}', [HoiDongController::class, 'destroy'])->name('admin.hoi-dong.destroy');
 });
